requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
pyasn1==0.6.1
PyJWT==2.9.0
python-jose==3.3.0
rest-framework-simplejwt==0.0.2
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
